requirements.txt
datasets==4.0.0
einops==0.8.1
flash-attn==2.5.8
fsspec==2025.3.0
huggingface_hub==0.27.1
hydra-core==1.3.2
lightning==2.5.2
numpy==1.26.4
omegaconf==2.3.0
Requests==2.32.4
rich==14.0.0
scipy==1.16.0
timm==1.0.17
tokenizers==0.21.2
torchmetrics==1.6.1
transformers==4.49.0

git-lfs==1.6
h5py==3.10.0
ipdb==0.13.13
notebook==7.1.1
nvitop==1.3.2
packaging==23.2
pandas
seaborn==0.13.2
scikit-learn
triton
torchaudio==2.3.1
torchvision==0.18.1
wandb
ocifs
hf_transfer